Updated search bar to render list of destinations
