+++ client/components/SearchBar.js
@@ -1,7 +1,8 @@
 import React, { useState } from 'react';
-import { Box, TextField } from '@mui/material';
+import { Box, List, ListItemButton, Paper, TextField, Typography } from '@mui/material';
 import { styled } from '@mui/material/styles';
-import { useNavigate } from 'react-router-dom';
+import { useDispatch, useSelector } from 'react-redux';
+import { fetchDestinations } from '../store';
 
 const StyledTextField = styled(TextField)({
   '& .MuiInputLabel-root': { color: 'white' },
@@ -25,17 +26,16 @@ const StyledTextField = styled(TextField)({
 });
 
 const SearchBar = () => {
-  const navigate = useNavigate();
+  const dispatch = useDispatch();
   const [searchValue, setSearchValue] = useState('');
+  const [renderDestinations, setRenderDestinations] = useState(false);
 
-  const handleSubmit = (event) => {
+  const destinations = useSelector((state) => state.destinations.destinations);
+
+  const handleSubmit = async (event) => {
     event.preventDefault();
-    navigate('/destinations', {
-      state: {
-        searchValue,
-      },
-    });
-    setSearchValue('');
+    await dispatch(fetchDestinations({ identifier: searchValue }));
+    setRenderDestinations(true);
   };
 
   return (
@@ -54,9 +54,44 @@ const SearchBar = () => {
         label="Where to"
         size={'small'}
         value={searchValue}
-        onChange={(event) => setSearchValue(event.target.value)}
+        onChange={(event) => {
+          setSearchValue(event.target.value);
+          setRenderDestinations(false);
+        }}
         inputProps={{ style: { color: '#fff' } }}
       />
+      {renderDestinations && destinations.length && <Paper style={{ position: 'absolute' }}>
+        <List component="nav" aria-label="main mailbox folders">
+          {destinations.map(city => <List key={city.id}>
+              <ListItemButton onClick={() => console.log('redirecting to city activities')}>
+                <Typography
+                  variant="h5"
+                  align="center"
+                  color="text.secondary"
+                >
+                  {city.name}
+                </Typography>
+              </ListItemButton>
+              <List component="nav">
+                {city.destinations.length && city.destinations.map(destination =>
+                  <ListItemButton key={destination.id}
+                                  style={{ paddingLeft: '2rem' }}
+                                  onClick={() => console.log('redirecting to destination activities')}>
+                    <Typography
+                      variant="h5"
+                      align="center"
+                      color="text.secondary"
+                      paragraph
+                    >
+                      {destination.name}
+                    </Typography>
+                  </ListItemButton>)}
+              </List>
+            </List>,
+          )}
+        </List>
+      </Paper>
+      }
     </Box>
   );
 };
