--- client/components/pages/destinations/index.js
@@ -1,3 +1,0 @@
-import DestinationsPage from './Destinations';
-
-export default DestinationsPage;
